Rework Support for Micrometer
Previously the support for Micrometer used a type that implied a complete
replacement of all other Micrometer functionality.  This was problematic as it
meant that any user customization wouldn't be taken into account.  This change
updates the configuration to use a type that causes our contribution to be
additive instead of a replacement.  Unfortunately this also means that we're
now aware of Spring Boot 1 and Spring Boot 2 compatibility differences and
we've had to break the implementations apart for conflicting version reasons.