Remove Required Properties
Previously a small set of properties were required as they always existed in a
CF LRP container.  However, when running a staged application as a task many
of these properties were not populated.  This change updates the
auto-configuration to remove the requirement that any of the properties
exists, making a best effort when they can be found.

Signed-off-by: Ben Hale <bhale@pivotal.io>